# server/docs/swagger.yaml
basePath: /api/fileshare/v1
definitions:
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      details: {}
      message:
        type: string
    type: object
info:
  contact:
    name: HaydenGuo
    url: https://hayden.pub
  description: File Share Manager workspace file sharing and governance API.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  termsOfService: https://github.com/ghl1024/file-share-manager
  title: File Share Manager API
  version: "1.0"
paths:
...
  /management/uploads/{id}:
    delete:
      consumes:
      - application/json
      description: Handles DELETE /api/fileshare/v1/management/uploads/{id}.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Set to XMLHttpRequest when using the session cookie
        in: header
        name: X-Requested-With
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Cancel
      tags:
      - Uploads
    get:
      consumes:
      - application/json
      description: Handles GET /api/fileshare/v1/management/uploads/{id} and GET /api/fileshare/v1/management/uploads/{id}/status.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Status
      tags:
      - Uploads
  /management/uploads/{id}/complete:
    post:
      consumes:
      - application/json
      description: Handles POST /api/fileshare/v1/management/uploads/{id}/complete.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          type: object
      - description: Set to XMLHttpRequest when using the session cookie
        in: header
        name: X-Requested-With
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Complete
      tags:
      - Uploads
  /management/uploads/{id}/parts/{part_no}:
    put:
      consumes:
      - application/octet-stream
      description: Handles PUT /api/fileshare/v1/management/uploads/{id}/parts/{part_no}.
      parameters:
      - description: Binary request body
        in: body
        name: body
        required: true
        schema:
          type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: part_no
        in: path
        name: part_no
        required: true
        type: string
      - description: Set to XMLHttpRequest when using the session cookie
        in: header
        name: X-Requested-With
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Part
      tags:
      - Uploads
  /management/uploads/{id}/status:
    get:
      consumes:
      - application/json
      description: Handles GET /api/fileshare/v1/management/uploads/{id} and GET /api/fileshare/v1/management/uploads/{id}/status.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Status
      tags:
      - Uploads
  /management/uploads/init:
    post:
      consumes:
      - application/json
      description: Handles POST /api/fileshare/v1/management/uploads/init.
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          type: object
      - description: Set to XMLHttpRequest when using the session cookie
        in: header
        name: X-Requested-With
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Init
      tags:
      - Uploads
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}". Browser sessions may also use the HttpOnly
      fileshare_session cookie.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"